pyproject.toml
[project]
name = "jsonrpcd"
version = "0.1.0"
description = ""
authors = [
    {name = "Mathieu Lecarme",email = "mathieu@garambrogne.net"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp (>=3.13.2,<4.0.0)",
    "aiodns (>=3.6.1,<4.0.0)",
    "aiohttp-sse (>=2.2.0,<3.0.0)",
    "aioconsole (>=0.8.2,<0.9.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
"pytest" = "^9"
ruff = "^0.14.10"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
isort = "^7.0.0"

